address test error on ci
